moved parsers to separate package and common functions to common package. fixed some tests also
