+++ api/fiat.go
@@ -5,12 +5,13 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"github.com/hermeznetwork/hermez-node/api/parsers"
 	"github.com/hermeznetwork/hermez-node/db/historydb"
 )
 
 func (a *API) getFiatCurrency(c *gin.Context) {
 	// Get symbol
-	symbol, err := parseCurrencyFilter(c)
+	symbol, err := parsers.ParseCurrencyFilter(c)
 	if err != nil {
 		retBadReq(errors.New(ErrInvalidSymbol), c)
 		return
@@ -26,7 +27,7 @@ func (a *API) getFiatCurrency(c *gin.Context) {
 
 func (a *API) getFiatCurrencies(c *gin.Context) {
 	// Currency filters
-	symbols, err := parseCurrenciesFilters(c)
+	symbols, err := parsers.ParseCurrenciesFilters(c)
 	if err != nil {
 		retBadReq(err, c)
 		return
